+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * CORE: [FIX] sort operation does not work for sorting data object, endpoint, collection resources
 * CORE: [FIX] Method `StreamIterator\StreamIterator::__toString()` must not throw an exception, caught ErrorException: Undefined index: created in
 * API: [CHANGE] If authentication failed a 401 gets returned instead of a 500
+* API: [CHANGE] filter is now a json encoded object in Process and Job resources
+* CORE: [FIX] skip garbage collection if a filtered process was issued
 
 
 ## 1.0.0-beta10